+++ public/kibana-integrations/kibana-vis.js
@@ -397,6 +397,7 @@ class KibanaVis extends Component {
         this.$rootScope.rendered = 'false';
       }
     }
+    if (typeof this.props.onRenderComplete === 'function') this.props.onRenderComplete();
   };
 
   showDateRangePicker = () => {