+++ test/cypress/cypress/integration/step-definitions/management/groups/the-confirm-modal-is-displayed-then.js
@@ -1,0 +1,8 @@
+import { Then } from 'cypress-cucumber-preprocessor/steps';
+import { elementIsVisible, getSelector } from '../../../utils/driver';
+import { GROUPS_PAGE as pageName} from '../../../utils/pages-constants';
+const confirmModalSelector = getSelector('confirmModalSelector', pageName);
+
+Then('The informative modal is displayed', () => {
+  elementIsVisible(confirmModalSelector);
+});