+++ test/cypress/cypress/integration/utils/mappers/wzd/reporting-modules-mapper.js
@@ -1,0 +1,7 @@
+import {
+  REPORTING_PAGE
+  } from '../../../pageobjects/wzd/reporting/report.page';
+
+export const REPORTING_PAGE_DATA = {
+    'generateReportButton': REPORTING_PAGE['generateReportButton']
+}